Refactor audio support, allow Cue redirect
CueSheets: Redirection of CueSheets is now done on startup instead of
when the CueSheet is loaded. Additionally, new CueSheets can be
injected into the game.

Cues: Individual Cues can now be redirected to another CueSheet. In
conjunction with injected a new CueSheet, this can be used to redirect
a single Cue in a CueSheet so that multiple recirects can update parts
of a CueSheet without interferring with each other.

Disclaimer: Unloading of injected CueSheets in not yet supported