Add a Ref Counter for injected Cue Sheets
It's possible to have Cues from multiple Cue Sheets redirected to the
same injected Cue Sheet. A Ref Counter was added to make sure that the
injected Cue Sheet isn't loaded multiple times as well as to make sure
it's only unloaded if every Cue Sheet referencing it has been unloaded
as well.